# api-docs.json
{"openapi":"3.0.1","info":{"title":"API de Carrito","description":"API para gestionar el carrito de compras","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/specialdates/validate":{"post":{"tags":["special-date-controller"],"operationId":"validateSpecialDate","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/login":{"post":{"tags":["login-controller"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenInfo"}}}}}}},"/cart/purchase":{"post":{"tags":["cart-controller"],"operationId":"processPurchase","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/vip/history":{"get":{"tags":["vip-history-controller"],"operationId":"getVipHistory","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VipHistoryDTO"}}}}}}}},"/user":{"get":{"tags":["user-controller"],"operationId":"getAuthenticatedUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/product/all":{"get":{"tags":["product-controller"],"operationId":"getAllProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}}}}}},"/cart/history":{"get":{"tags":["cart-controller"],"operationId":"getPurchaseHistory","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"AuthenticationReq":{"type":"object","properties":{"usuario":{"type":"string"},"clave":{"type":"string"}}},"TokenInfo":{"type":"object","properties":{"jwtToken":{"type":"string"}}},"CartDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"creationDate":{"type":"string","format":"date-time"},"total":{"type":"number"},"cartType":{"type":"string","enum":["VIP","COMMON","SPECIAL_DATE"]},"user":{"$ref":"#/components/schemas/UserDTO"},"cartDetails":{"type":"array","items":{"$ref":"#/components/schemas/CartDetailDTO"}}}},"CartDetailDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"product":{"$ref":"#/components/schemas/ProductDTO2"},"amount":{"type":"integer","format":"int32"},"totalPrice":{"type":"number"}}},"ProductDTO2":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"number"}}},"UserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"rol":{"type":"string","enum":["USER","ADMIN","VIP"]},"lastBuy":{"type":"string","format":"date-time"},"voucher":{"type":"boolean"},"email":{"type":"string"}}},"VipHistoryDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"startVip":{"type":"string","format":"date-time"},"endVip":{"type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/UserDTO"}}},"ProductDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nameProduct":{"type":"string"},"price":{"type":"number"},"stock":{"type":"integer","format":"int32"},"imageUrl":{"type":"string"}}}}}}