# carrito/src/main/java/com/example/carrito/DTO/ProductDTO.java
package com.example.carrito.DTO;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.math.BigDecimal;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class ProductDTO {

    private Long id;
    private String nameProduct;
    private BigDecimal price;
    private int stock;
    private String imageUrl;
}